pycrypto/requirements.txt
pycrypto==2.6.1.
